Move the static analysis profile data structs to the api package, so we can use them outside of the vulkan package.
