+++ gapis/api/vulkan/profile_static_analysis.go
@@ -28,31 +28,6 @@ import (
 	"github.com/google/gapid/gapis/service/path"
 )
 
-// StaticAnalysisProfileData is the result of the profiling static analysis.
-// TODO(pmuetschard): this is a temporary struct here, that will be replaced once the
-// service.ProfilingData is cleaned up/refactored for the new PerfTab.
-type StaticAnalysisProfileData struct {
-	CounterSpecs []CounterSpec
-	CounterData  []CounterData
-}
-
-type CounterSpec struct {
-	ID          uint32
-	Name        string
-	Description string
-	Unit        string // this should match the unit from the Perfetto data.
-}
-
-type Sample struct {
-	Counter uint32
-	Value   float64
-}
-
-type CounterData struct {
-	Index   api.SubCmdIdx
-	Samples []Sample
-}
-
 type counterType uint32
 
 const (
@@ -64,7 +39,7 @@ const (
 )
 
 var (
-	counters = map[counterType]CounterSpec{
+	counters = map[counterType]api.StaticAnalysisCounter{
 		vertexCountType: {
 			ID:          uint32(vertexCountType),
 			Name:        "Vertex Count",
@@ -99,21 +74,21 @@ var (
 )
 
 type profilingDataBuilder struct {
-	data         *StaticAnalysisProfileData
+	data         *api.StaticAnalysisProfileData
 	seenCounters map[counterType]struct{}
 }
 
-func (db profilingDataBuilder) newSampler(idx api.SubCmdIdx) *CounterData {
-	db.data.CounterData = append(db.data.CounterData, CounterData{Index: idx})
+func (db profilingDataBuilder) newSampler(idx api.SubCmdIdx) *api.StaticAnalysisCounterSamples {
+	db.data.CounterData = append(db.data.CounterData, api.StaticAnalysisCounterSamples{Index: idx})
 	return &db.data.CounterData[len(db.data.CounterData)-1]
 }
 
-func (db profilingDataBuilder) addSample(sampler *CounterData, counter counterType, value float64) {
+func (db profilingDataBuilder) addSample(sampler *api.StaticAnalysisCounterSamples, counter counterType, value float64) {
 	if _, ok := db.seenCounters[counter]; !ok {
 		db.seenCounters[counter] = struct{}{}
 		db.data.CounterSpecs = append(db.data.CounterSpecs, counters[counter])
 	}
-	sampler.Samples = append(sampler.Samples, Sample{uint32(counter), value})
+	sampler.Samples = append(sampler.Samples, api.StaticAnalysisCounterSample{uint32(counter), value})
 }
 
 // Construct a decoder for reading buffer memory
@@ -365,9 +340,9 @@ func getIndirectCount(ctx context.Context, s *api.GlobalState, countBuffer VkBuf
 // ProfileStaticAnalysis computes the static analysis profiling data. It processes each command
 // and for each submitted draw call command, it computes statistics to be shown as counters in
 // PerfTab, combined with the hardware counters.
-func (API) ProfileStaticAnalysis(ctx context.Context, p *path.Capture) (interface{}, error) {
+func (API) ProfileStaticAnalysis(ctx context.Context, p *path.Capture) (*api.StaticAnalysisProfileData, error) {
 	data := profilingDataBuilder{
-		data:         &StaticAnalysisProfileData{},
+		data:         &api.StaticAnalysisProfileData{},
 		seenCounters: map[counterType]struct{}{},
 	}
 	postSubCmdCb := func(s *api.GlobalState, idx api.SubCmdIdx, cmd api.Cmd, ref interface{}) {
